backend/requirements.txt
asgiref==3.4.1
backcall==0.2.0
bidict==0.21.4
decorator==5.1.0
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.8.2
ipython==7.29.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.3.0
pyotp==2.6.0
python-engineio==4.3.0
python-socketio==5.4.1
pytz==2021.3
sqlparse==0.4.2
traitlets==5.1.1
typing-extensions==3.10.0.2
wcwidth==0.2.5
whitenoise==5.3.0
zipp==3.6.0
